location get user input

+++ types/location.go
@@ -3,12 +3,12 @@ package types
 type LocationResponse struct {
 	EventID   int     `json:"eventID"`
 	UserID    int     `json:"userID"`
-	Latitude  float32 `json:"latitude"`
-	Longitude float32 `json:"longitude"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 	Address   string  `json:"address"`
 }
 
 type Coordinates struct {
-	Latitude  float32 `json:"latitude"`
-	Longitude float32 `json:"longitude"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 }
